// js-lib/demo/src/utils/yakuName.d.ts
export declare const yakuName: {
    Tsumo: string;
    Pinhu: string;
    Tanyao: string;
    Ipe: string;
    SelfWind: string;
    RoundWind: string;
    Haku: string;
    Hatsu: string;
    Chun: string;
    Sanshoku: string;
    Ittsu: string;
    Chanta: string;
    Chitoi: string;
    Toitoi: string;
    Sananko: string;
    Honroto: string;
    Sandoko: string;
    Sankantsu: string;
    Shosangen: string;
    Honitsu: string;
    Junchan: string;
    Ryanpe: string;
    Chinitsu: string;
    Richi: string;
    Ippatsu: string;
    Rinshan: string;
    Chankan: string;
    Haitei: string;
    Houtei: string;
    WRichi: string;
    Tenhou: string;
    Chihou: string;
    Kokushi: string;
    Suanko: string;
    Daisangen: string;
    Tsuiso: string;
    Shousushi: string;
    Lyuiso: string;
    Chinroto: string;
    Sukantsu: string;
    Churen: string;
    Daisushi: string;
    ChurenNineWaiting: string;
    SuankoTanki: string;
    KokushiThirteenWaiting: string;
};
export declare const doubleTimesYakuman: string[];